Remove livesync-provider
Remove livesyncProvider and introduce a base class for device specific livesync services.
